trello_functions/requirements.txt
azure_storage==0.37.0
py_trello==0.19.0
python-dotenv==1.0.0
